move setupOffscreenDocumet to offscreen.ts
