+++ src/tinyPages/offscreen.ts
@@ -37,6 +37,37 @@ export type RecordErrorMessage = {
   };
 };
 
+let creating: Promise<void> | null = null;
+export async function setupOffscreenDocument(path: string) {
+  const offscreenUrl = chrome.runtime.getURL(path);
+  const existingContexts = await chrome.runtime.getContexts({
+    // @ts-expect-error -- TODO the type seems to be wrong here?
+    contextTypes: ["OFFSCREEN_DOCUMENT"],
+    documentUrls: [offscreenUrl],
+  });
+
+  // @ts-expect-error -- TODO type contradicts the chrome api docs?
+  if (existingContexts.length > 0) {
+    return;
+  }
+
+  if (creating == null) {
+    creating = chrome.offscreen.createDocument({
+      url: "offscreen.html",
+      // Our reason for creating an offscreen document does not fit nicely into options offered by the Chrome API, which
+      // is error telemetry. Other possible options: TESTING or WORKERS. We chose BLOBS because it's the closest to
+      // interaction with error objects?
+      reasons: [chrome.offscreen.Reason.BLOBS],
+      justification:
+        "Error telemetry SDK usage that is incompatible with service workers",
+    });
+    await creating;
+    creating = null;
+  } else {
+    await creating;
+  }
+}
+
 function isRecordErrorMessage(message: unknown): message is RecordErrorMessage {
   if (typeof message !== "object" || message == null) {
     return false;
